divided init_light_bonus.c into two files
